Refactor to module to preserve genesis config generation code for posterity
